+++ src/components/Home/LinkedImgSection/LinkedImgSection.js
@@ -2,8 +2,6 @@ import React from 'react';
 import './LinkedImgSection.css';
 import LinkedImageCard from '../../ImageCards/LinkedImageCard';
 import imagesInfo from './data';
-import imageInfo from '../LeftSliderSection/data';
-
 
 function LinkedImgSection() {
     return (
