+++ src/components/Home/ProductSection/Products.css
@@ -38,96 +38,7 @@
     padding: 60px;
   }
   
-  .main_product_card {
-    position: relative;
-    width: 20%;
-    margin: 10px 10px 10px 0;
-    padding-top: 21%;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-        -ms-flex-direction: column;
-            flex-direction: column;
-            align-items: flex-start;
-  }
-  
-  .main_product_card_container {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    left: 0;
-  }
-  
-  .main_product_image_container img {
-    width: 100%;
-    height: 100%;
-  }
-  
-  .main_product_add_btn {
-    width: 60%;
-    height: 40px;
-    background: #cea492;
-    color: #fff;
-    text-transform: capitalize;
-    position: absolute;
-    left: 50%;
-    top:55%;
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-pack: center;
-        -ms-flex-pack: center;
-            justify-content: center;
-    -webkit-box-align: center;
-        -ms-flex-align: center;
-            align-items: center;
-    -webkit-transform: translateX(-50%);
-            transform: translateX(-50%);
-    cursor: pointer;
-    opacity: 0;
-    -webkit-transition: all .3s ease;
-    transition: all .3s ease;
-  }
-  
-  .main_product_add_btn:hover {
-    background: #000;
-  }
-  
-  .main_product_card:hover .main_product_add_btn {
-    opacity: 1;
-  }
-  
-  .main_card_name {
-    color: #000;
-  }
-  
-  @media screen and (max-width: 1200px) {
-    .main_page_container {
-      width: 85%;
-    }
-    .main_products_title h3 {
-      font-size: 30px;
-    }
-  }
-  
-  @media screen and (max-width: 992px) {
-    .main_page_container {
-      width: 95%;
-    }
-    .main_product_card {
-      width: 45%;
-      padding-top: 46%;
-    }
-  }
-  
   @media screen and (max-width: 500px) {
-    .main_product_card {
-      width: 70%;
-      padding-top: 70%;
-    }
     .main_product_list {
       -webkit-box-pack: center;
           -ms-flex-pack: center;
