+++ src/components/Home/ProductSection/Products.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import './Products.css';
+import ProductCard from '../../ProductCard/ProductCard';
 
 function Products({title, productList}){
     return (
@@ -11,21 +12,7 @@ function Products({title, productList}){
                   </h3>
               </div>
               <div className="main_product_list">
-                {productList.map(product =>{
-                    const {id, name, image, price} = product;
-                    return (
-                      <article key={id} className="main_product_card">
-                        <a href="#" className="main_product_card_container">
-                            <div className="main_product_image_container">
-                                <img src={image} alt={name} />
-                            </div>
-                        </a>
-                        <a href="#" className="main_product_add_btn white-text"> Add to card</a>
-                        <a className="main_card_name" href="#">{name.length <= 32 ? name : name.slice(0,32) + '...'}</a>
-                        <p><span>$</span> <span>{price}</span></p>
-                    </article>
-                    )
-                })}
+                {productList.map(product => <ProductCard {...product} />)}
               </div>
           </div>
       </section>
